pyproject.toml
[tool.poetry]
name = "invest-contest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
